# [POAE7-2809]try to optimize expression performance