--- a/cpp/src/cider/exec/module/CiderOptions.cpp
+++ b/cpp/src/cider/exec/module/CiderOptions.cpp
@@ -51,3 +51,9 @@ DEFINE_bool(allow_runtime_query_interrupt, false, "allow runtime query interrupt
 DEFINE_double(running_query_interrupt_freq, 0.5, "running query interrupt freq");
 DEFINE_uint64(pending_query_interrupt_freq, 1000, "pending query interrupt freq");
 DEFINE_bool(force_direct_hash, false, "force direct hash");
+
+// Codegen Options
+DEFINE_bool(codegen_all_opt, false, "codegen all optimization");
+DEFINE_bool(check_bit_vector_clear_opt, false, "use optimized bit clear checker");
+DEFINE_bool(set_null_bit_vector_opt, false, "use optimized set null bit");
+DEFINE_bool(branchless_logic, false, "use branchless logic operation");