--- a/cpp/src/cider/exec/nextgen/jitlib/llvmjit/LLVMJITValue.cpp
+++ b/cpp/src/cider/exec/nextgen/jitlib/llvmjit/LLVMJITValue.cpp
@@ -31,7 +31,6 @@
 namespace cider::jitlib {
 JITValue& LLVMJITValue::assign(JITValue& value) {
   if (!is_variable_) {
-    // return *this;
     LOG(FATAL) << "JITValue " << getValueName()
                << "is not a variable in LLVMJITValue::assign.";
   }
@@ -70,6 +69,9 @@ JITValuePointer LLVMJITValue::orOp(JITValue& rh) {
       ans = getFunctionBuilder(parent_function_).CreateOr(load(), llvm_rh.load());
       break;
     }
+    case JITTypeTag::INT8:
+      ans = getFunctionBuilder(parent_function_).CreateOr(load(), llvm_rh.load());
+      break;
     default:
       LOG(FATAL) << "Invalid JITValue type for or operation. Name=" << getValueName()
                  << ", Type=" << getJITTypeName(getValueTypeTag()) << ".";