--- a/cpp/src/cider/exec/nextgen/operators/ProjectNode.h
+++ b/cpp/src/cider/exec/nextgen/operators/ProjectNode.h
@@ -41,9 +41,11 @@ class ProjectTranslator : public Translator {
   using Translator::Translator;
 
   void consume(context::CodegenContext& context) override;
+  void consumeNull(context::CodegenContext& context) override;
 
  private:
   void codegen(context::CodegenContext& context);
+  void codegenNull(context::CodegenContext& context);
 };
 }  // namespace cider::exec::nextgen::operators
 #endif  // NEXTGEN_OPERATORS_PROJECTNODE_H