--- a/cpp/src/cider/exec/nextgen/operators/RowToColumnNode.cpp
+++ b/cpp/src/cider/exec/nextgen/operators/RowToColumnNode.cpp
@@ -43,7 +43,7 @@ class ColumnWriter {
       , index_(index)
       , arrow_array_len_(arrow_array_len) {}
 
-  void write() {
+  void write(bool for_null) {
     // TBD: Whether input ColumnVar's ArrowArray could be overwriten.
     // CHECK(0 == expr_->getLocalIndex());
     switch (expr_->get_type_info().get_type()) {
@@ -57,12 +57,12 @@ class ColumnWriter {
       case kDATE:
       case kTIMESTAMP:
       case kTIME:
-        writeFixSizedTypeCol();
+        writeFixSizedTypeCol(for_null);
         break;
       case kVARCHAR:
       case kCHAR:
       case kTEXT:
-        writeVariableSizeTypeCol();
+        writeVariableSizeTypeCol(for_null);
         break;
       default:
         LOG(FATAL) << "Unsupported data type in ColumnWriter: "
@@ -71,9 +71,17 @@ class ColumnWriter {
   }
 
  private:
-  void writeVariableSizeTypeCol() {
+  void writeVariableSizeTypeCol(bool for_null) {
     // Get values need to write
     utils::VarSizeJITExprValue values(expr_->get_expr_value());
+
+    if (for_null) {
+      auto null = setNullBuffer(values.getNull(), for_null);
+      Analyzer::JITExprValueAdaptor(
+          context_.getArrowArrayValues(expr_->getLocalIndex()).second)
+          .setNull(null);
+      return;
+    }
 
     // Allocate buffer
     // offset, need array_len + 1 element.
@@ -110,15 +118,23 @@ class ColumnWriter {
     size_t local_offset = context_.appendArrowArrayValues(
         arrow_array_,
         utils::JITExprValue(JITExprValueType::BATCH,
-                            setNullBuffer(values.getNull()),
+                            setNullBuffer(values.getNull(), for_null),
                             raw_length_buffer,
                             raw_data_buffer));
     expr_->setLocalIndex(local_offset);
   }
 
-  void writeFixSizedTypeCol() {
+  void writeFixSizedTypeCol(bool for_null) {
     // Get values need to write
     utils::FixSizeJITExprValue values(expr_->get_expr_value());
+
+    if (for_null) {
+      auto null = setNullBuffer(values.getNull(), for_null);
+      Analyzer::JITExprValueAdaptor(
+          context_.getArrowArrayValues(expr_->getLocalIndex()).second)
+          .setNull(null);
+      return;
+    }
 
     // Allocate buffer.
     // Write value
@@ -126,8 +142,9 @@ class ColumnWriter {
     // Register Output ArrowArray to CodegenContext
     size_t local_offset = context_.appendArrowArrayValues(
         arrow_array_,
-        utils::JITExprValue(
-            JITExprValueType::BATCH, setNullBuffer(values.getNull()), raw_data_buffer));
+        utils::JITExprValue(JITExprValueType::BATCH,
+                            setNullBuffer(values.getNull(), for_null),
+                            raw_data_buffer));
     expr_->setLocalIndex(local_offset);
   }
 
@@ -171,18 +188,22 @@ class ColumnWriter {
     }
   }
 
-  JITValuePointer setNullBuffer(JITValuePointer& null_val) {
+  JITValuePointer setNullBuffer(JITValuePointer& null_val, bool for_null) {
     // the null_buffer, raw_data_buffer not used anymore.
     // so it doesn't matter whether null_buffer is nullptr
     // or constant false.
     auto null_buffer = JITValuePointer(nullptr);
-    if (!FLAGS_null_separate && !expr_->get_type_info().get_notnull()) {
+    if ((!FLAGS_null_separate || for_null) && !expr_->get_type_info().get_notnull()) {
       // auto tmp_8bit_acc =
       //     context_.getJITFunction()->createVariable(JITTypeTag::INT8, "tmp_8bit_acc",
       //     0);
       // TBD: Null representation, bit-array or bool-array.
       null_buffer.replace(context_.getJITFunction()->createLocalJITValue(
           [this]() { return allocateBitwiseBuffer(0); }));
+      if (for_null) {
+        *null_buffer[index_] = null_val;
+        return null_buffer;
+      }
 
       std::string fname = "set_null_vector_bit";
       if (context_.getCodegenOptions().set_null_bit_vector_opt) {
@@ -235,25 +256,31 @@ void RowToColumnTranslator::consume(context::CodegenContext& context) {
   codegen(context);
 }
 
-void RowToColumnTranslator::codegen(context::CodegenContext& context) {
+void RowToColumnTranslator::consumeNull(context::CodegenContext& context) {
+  codegen(context, true);
+}
+
+void RowToColumnTranslator::codegen(context::CodegenContext& context, bool for_null) {
   auto func = context.getJITFunction();
   auto&& [type, exprs] = node_->getOutputExprs();
   ExprPtrVector& output_exprs = exprs;
 
   // construct output Batch
-  auto output_arrow_array = context.registerBatch(
-      SQLTypeInfo(kSTRUCT,
-                  false,
-                  [&output_exprs]() {
-                    std::vector<SQLTypeInfo> output_types;
-                    output_types.reserve(output_exprs.size());
-                    for (auto& expr : output_exprs) {
-                      output_types.emplace_back(expr->get_type_info());
-                    }
-                    return output_types;
-                  }()),
-      "output",
-      true);
+  if (!for_null) {
+    output_arrow_array_.replace(context.registerBatch(
+        SQLTypeInfo(kSTRUCT,
+                    false,
+                    [&output_exprs]() {
+                      std::vector<SQLTypeInfo> output_types;
+                      output_types.reserve(output_exprs.size());
+                      for (auto& expr : output_exprs) {
+                        output_types.emplace_back(expr->get_type_info());
+                      }
+                      return output_types;
+                    }()),
+        "output",
+        true));
+  }
 
   auto output_index = func->createVariable(JITTypeTag::INT64, "output_index", 0);
 
@@ -264,21 +291,28 @@ void RowToColumnTranslator::codegen(context::CodegenContext& context) {
   for (int64_t i = 0; i < exprs.size(); ++i) {
     ExprPtr& expr = exprs[i];
 
-    auto child_arrow_array = func->createLocalJITValue([&output_arrow_array, &i]() {
-      return codegen_utils::getArrowArrayChild(output_arrow_array, i);
+    auto child_arrow_array = func->createLocalJITValue([&i, this]() {
+      return codegen_utils::getArrowArrayChild(output_arrow_array_, i);
     });
 
     // Write rows
     ColumnWriter writer(context, expr, child_arrow_array, output_index, input_array_len);
-    writer.write();
+    writer.write(for_null);
   }
   // Update index
   output_index = output_index + 1;
 
+  if (for_null) {
+    if (successor_) {
+      successor_->consumeNull(context);
+    }
+    return;
+  }
+
   // Execute length field updating build function after C2R loop finished.
   prev_c2r_node->registerDeferFunc(
-      [output_index, output_arrow_array, &output_exprs, &context]() mutable {
-        codegen_utils::setArrowArrayLength(output_arrow_array, output_index);
+      [output_index, &output_exprs, &context, this]() mutable {
+        codegen_utils::setArrowArrayLength(output_arrow_array_, output_index);
         for (auto& expr : output_exprs) {
           size_t local_offset = expr->getLocalIndex();
           CHECK_NE(local_offset, 0);
@@ -292,4 +326,5 @@ void RowToColumnTranslator::codegen(context::CodegenContext& context) {
     successor_->consume(context);
   }
 }
+
 }  // namespace cider::exec::nextgen::operators