… cpp/src/cider-velox/benchmark/expression/ArithmeticAndComparison.cpp
… cpp/src/cider/exec/plan/parser/TypeUtils.h
… cpp/src/cider/exec/processor/DefaultBatchProcessor.cpp
… cpp/src/cider/exec/processor/DefaultBatchProcessor.h
… cpp/src/cider/function/scalar/RuntimeFunctions.cpp
… cpp/src/cider/include/cider/CiderOptions.h
… cpp/src/cider/type/plan/BinaryExpr.cpp
… cpp/src/cider/type/plan/BinaryExpr.h
… cpp/src/cider/type/plan/ColumnExpr.cpp
… cpp/src/cider/type/plan/ColumnExpr.h
… cpp/src/cider/type/plan/ConstantExpr.cpp
… cpp/src/cider/type/plan/ConstantExpr.h
… cpp/src/cider/type/plan/Expr.cpp
… cpp/src/cider/type/plan/Expr.h
… cpp/src/cider/util/CiderBitUtils.h