### What changes were proposed in this pull request?
try to optimize expression performance
1) branchless null set and logic operation
2) xor replace not(and) for bit clear check
3) null separated for auto-vectorization
all optimization is under gflags, so nothing changed if all gflags are closed.

**branchless logic gain 15%~35% performance boost for logic operation**

**auto-vectorization will gain 2x performance boost, but null process is hard**
**this pr just enable auto-vectorization with null process splited, which hard to support keleene logic operation**
**so we propagate null, which is one strategy in arrow kernel**

some performance details on haswell:

```bash
============================================================================
/workspace/code/BDTK/src/cider-velox/benchmark/expression/ArithmeticAndComparison.cpprelative  time/iter  iters/s
============================================================================
velox                                                         3.75s  266.78m
nextgen                                           90.84%      4.13s  242.34m
nextgen_opt                                      228.17%      1.64s  608.69m
specifialize1                                   1923.43%   194.89ms     5.13
specifialize2                                   1146.42%   326.97ms     3.06
specifialize3                                    241.53%      1.55s  644.35m
----------------------------------------------------------------------------
mulI8Velox_________Base                                       3.79s  263.63m
mulI8NextGen                                      92.60%      4.10s  244.13m
mulI8NextGen_______Base                                       4.10s  243.78m
mulI8NextgenIsBitClear                            93.25%      4.40s  227.33m
mulI8NextgenBranchlessSetNull                     38.49%     10.66s   93.82m
mulI8NextgenBranchlessLogic                       98.83%      4.15s  240.94m
mulI8NextgenNullSeparate                         292.02%      1.40s  711.91m
mulI8NextgenAllOpt                               296.91%      1.38s  723.81m
mulI8Compile                                                57.50ms    17.39
----------------------------------------------------------------------------
mulI8NestedVelox_________Base                                 6.20s  161.20m
mulI8NestedNextGen                               144.74%      4.29s  233.32m
mulI8NestedNextGen_______Base                                 4.38s  228.28m
mulI8NestedNextgenIsBitClear                      96.13%      4.56s  219.45m
mulI8NestedNextgenBranchlessSetNull               84.12%      5.21s  192.04m
mulI8NestedNextgenBranchlessLogic                101.13%      4.33s  230.85m
mulI8NestedNextgenNullSeparate                   328.52%      1.33s  749.94m
mulI8NestedNextgenAllOpt                         327.48%      1.34s  747.57m
mulI8NestedCompile                                          58.75ms    17.02
----------------------------------------------------------------------------
mulI8NestedDeepVelox_________Base                             8.78s  113.90m
mulI8NestedDeepNextGen                           195.45%      4.49s  222.62m
mulI8NestedDeepNextGen_______Base                             4.44s  224.98m
mulI8NestedDeepNextgenIsBitClear                  97.28%      4.57s  218.86m
mulI8NestedDeepNextgenBranchlessSetNull           78.33%      5.67s  176.23m
mulI8NestedDeepNextgenBranchlessLogic             98.15%      4.53s  220.82m
mulI8NestedDeepNextgenNullSeparate               329.11%      1.35s  740.45m
mulI8NestedDeepNextgenAllOpt                     331.00%      1.34s  744.69m
mulI8NestedDeepCompile                                      57.64ms    17.35
----------------------------------------------------------------------------
mulI16Velox_________Base                                      3.14s  318.00m
mulI16NextGen                                     76.29%      4.12s  242.60m
mulI16NextGen_______Base                                      4.14s  241.65m
mulI16NextgenIsBitClear                           94.24%      4.39s  227.73m
mulI16NextgenBranchlessSetNull                    38.99%     10.61s   94.21m
mulI16NextgenBranchlessLogic                     100.41%      4.12s  242.63m
mulI16NextgenNullSeparate                        282.30%      1.47s  682.16m
mulI16NextgenAllOpt                              281.79%      1.47s  680.93m
mulI16Compile                                               58.10ms    17.21
----------------------------------------------------------------------------
mulI32Velox_________Base                                      3.14s  318.60m
mulI32NextGen                                     72.70%      4.32s  231.63m
mulI32NextGen_______Base                                      4.29s  233.13m
mulI32NextgenIsBitClear                           94.43%      4.54s  220.13m
mulI32NextgenBranchlessSetNull                    40.46%     10.60s   94.32m
mulI32NextgenBranchlessLogic                     101.08%      4.24s  235.64m
mulI32NextgenNullSeparate                        274.58%      1.56s  640.13m
mulI32NextgenAllOpt                              276.19%      1.55s  643.88m
mulI32Compile                                               59.38ms    16.84
----------------------------------------------------------------------------
mulI64Velox_________Base                                      3.04s  328.53m
mulI64NextGen                                     72.30%      4.21s  237.52m
mulI64NextGen_______Base                                      4.34s  230.30m
mulI64NextgenIsBitClear                           90.27%      4.81s  207.90m
mulI64NextgenBranchlessSetNull                    39.94%     10.87s   91.97m
mulI64NextgenBranchlessLogic                      99.67%      4.36s  229.54m
mulI64NextgenNullSeparate                        197.20%      2.20s  454.14m
mulI64NextgenAllOpt                              194.76%      2.23s  448.54m
mulI64Compile                                               59.23ms    16.88
----------------------------------------------------------------------------
mulDoubleVelox_________Base                                   2.97s  337.11m
mulDoubleNextGen                                  59.20%      5.01s  199.58m
mulDoubleNextGen_______Base                                   4.95s  201.94m
mulDoubleNextgenIsBitClear                        83.25%      5.95s  168.11m
mulDoubleNextgenBranchlessSetNull                 56.08%      8.83s  113.26m
mulDoubleNextgenBranchlessLogic                   99.48%      4.98s  200.89m
mulDoubleNextgenNullSeparate                     267.06%      1.85s  539.31m
mulDoubleNextgenAllOpt                           272.94%      1.81s  551.18m
mulDoubleCompile                                            60.50ms    16.53
----------------------------------------------------------------------------
mulDoubleSameColumnVelox_________Base                         3.29s  303.83m
mulDoubleSameColumnNextGen                        77.69%      4.24s  236.05m
mulDoubleSameColumnNextGen_______Base                         4.22s  237.01m
mulDoubleSameColumnNextgenIsBitClear              90.56%      4.66s  214.63m
mulDoubleSameColumnNextgenBranchlessSetNull       81.37%      5.19s  192.86m
mulDoubleSameColumnNextgenBranchlessLogic         97.71%      4.32s  231.58m
mulDoubleSameColumnNextgenNullSeparate           216.80%      1.95s  513.84m
mulDoubleSameColumnNextgenAllOpt                 213.21%      1.98s  505.34m
mulDoubleSameColumnCompile                                  59.88ms    16.70
----------------------------------------------------------------------------
mulDoubleConstantVelox_________Base                           2.99s  334.84m
mulDoubleConstantNextGen                          60.18%      4.96s  201.50m
mulDoubleConstantNextGen_______Base                           4.78s  209.26m
mulDoubleConstantNextgenIsBitClear                80.91%      5.91s  169.32m
mulDoubleConstantNextgenBranchlessSetNull         53.95%      8.86s  112.90m
mulDoubleConstantNextgenBranchlessLogic           97.82%      4.89s  204.69m
mulDoubleConstantNextgenNullSeparate             228.67%      2.09s  478.51m
mulDoubleConstantNextgenAllOpt                   207.58%      2.30s  434.38m
mulDoubleConstantCompile                                    60.90ms    16.42
----------------------------------------------------------------------------
mulDoubleNestedVelox_________Base                             4.48s  223.41m
mulDoubleNestedNextGen                            88.41%      5.06s  197.52m
mulDoubleNestedNextGen_______Base                             5.20s  192.16m
mulDoubleNestedNextgenIsBitClear                  88.63%      5.87s  170.32m
mulDoubleNestedNextgenBranchlessSetNull           79.82%      6.52s  153.37m
mulDoubleNestedNextgenBranchlessLogic            100.86%      5.16s  193.80m
mulDoubleNestedNextgenNullSeparate               229.01%      2.27s  440.07m
mulDoubleNestedNextgenAllOpt                     247.84%      2.10s  476.25m
mulDoubleNestedCompile                                      61.80ms    16.18
----------------------------------------------------------------------------
mulDoubleNestedDeepVelox_________Base                         9.02s  110.83m
mulDoubleNestedDeepNextGen                       182.55%      4.94s  202.31m
mulDoubleNestedDeepNextGen_______Base                         4.96s  201.68m
mulDoubleNestedDeepNextgenIsBitClear              83.17%      5.96s  167.73m
mulDoubleNestedDeepNextgenBranchlessSetNull       69.95%      7.09s  141.08m
mulDoubleNestedDeepNextgenBranchlessLogic         99.47%      4.99s  200.60m
mulDoubleNestedDeepNextgenNullSeparate           233.13%      2.13s  470.17m
mulDoubleNestedDeepNextgenAllOpt                 252.74%      1.96s  509.71m
mulDoubleNestedDeepCompile                                  62.76ms    15.93
----------------------------------------------------------------------------
PlusCheckedVeloxI64                                           3.96s  252.70m
plusI64Velox_________Base                                     3.34s  298.99m
plusI64NextGen                                    75.13%      4.45s  224.64m
plusI64NextGen_______Base                                     4.20s  238.11m
plusI64NextgenIsBitClear                          90.96%      4.62s  216.59m
plusI64NextgenBranchlessSetNull                   38.28%     10.97s   91.15m
plusI64NextgenBranchlessLogic                     98.23%      4.28s  233.90m
plusI64NextgenNullSeparate                       229.01%      1.83s  545.29m
plusI64NextgenAllOpt                             227.79%      1.84s  542.40m
plusI64Compile                                              61.27ms    16.32
----------------------------------------------------------------------------
mulAndAddVelox_________Base                                  16.24s   61.59m
mulAndAddNextGen                                 309.64%      5.24s  190.70m
mulAndAddNextGen_______Base                                   5.12s  195.28m
mulAndAddNextgenIsBitClear                        96.30%      5.32s  188.06m
mulAndAddNextgenBranchlessSetNull                 80.68%      6.35s  157.56m
mulAndAddNextgenBranchlessLogic                   97.58%      5.25s  190.56m
mulAndAddNextgenNullSeparate                     222.06%      2.31s  433.64m
mulAndAddNextgenAllOpt                           235.69%      2.17s  460.26m
mulAndAddCompile                                            59.19ms    16.90
----------------------------------------------------------------------------
eqVelox_________Base                                          3.83s  260.81m
eqNextGen                                         69.54%      5.51s  181.38m
eqNextGen_______Base                                          5.56s  179.81m
eqNextgenIsBitClear                               88.95%      6.25s  159.93m
eqNextgenBranchlessSetNull                        48.03%     11.58s   86.37m
eqNextgenBranchlessLogic                         103.14%      5.39s  185.46m
eqNextgenNullSeparate                            159.73%      3.48s  287.20m
eqNextgenAllOpt                                  115.42%      4.82s  207.53m
eqCompile                                                   62.35ms    16.04
----------------------------------------------------------------------------
eqConstantVelox_________Base                                  3.91s  255.52m
eqConstantNextGen                                 69.55%      5.63s  177.72m
eqConstantNextGen_______Base                                  5.60s  178.51m
eqConstantNextgenIsBitClear                       88.90%      6.30s  158.69m
eqConstantNextgenBranchlessSetNull                46.97%     11.93s   83.84m
eqConstantNextgenBranchlessLogic                  98.33%      5.70s  175.53m
eqConstantNextgenNullSeparate                    159.55%      3.51s  284.81m
eqConstantNextgenAllOpt                          115.50%      4.85s  206.17m
eqConstantCompile                                           60.47ms    16.54
----------------------------------------------------------------------------
eqBoolVelox_________Base                                      4.43s  225.99m
eqBoolNextGen                                     72.80%      6.08s  164.51m
eqBoolNextGen_______Base                                      6.02s  166.03m
eqBoolNextgenIsBitClear                           72.45%      8.31s  120.29m
eqBoolNextgenBranchlessSetNull                    68.81%      8.75s  114.25m
eqBoolNextgenBranchlessLogic                      99.58%      6.05s  165.33m
eqBoolNextgenNullSeparate                        132.70%      4.54s  220.32m
eqBoolNextgenAllOpt                               62.65%      9.61s  104.02m
eqBoolCompile                                               63.49ms    15.75
----------------------------------------------------------------------------
neqVelox_________Base                                         3.83s  261.14m
neqNextGen                                        68.61%      5.58s  179.18m
neqNextGen_______Base                                         5.57s  179.41m
neqNextgenIsBitClear                              82.74%      6.74s  148.44m
neqNextgenBranchlessSetNull                       48.84%     11.41s   87.62m
neqNextgenBranchlessLogic                        100.31%      5.56s  179.98m
neqNextgenNullSeparate                           161.81%      3.44s  290.31m
neqNextgenAllOpt                                 116.75%      4.77s  209.47m
neqCompile                                                  62.09ms    16.10
----------------------------------------------------------------------------
gtVelox_________Base                                          3.92s  255.04m
gtNextGen                                         68.74%      5.70s  175.32m
gtNextGen_______Base                                          5.71s  175.25m
gtNextgenIsBitClear                               82.97%      6.88s  145.41m
gtNextgenBranchlessSetNull                        71.14%      8.02s  124.68m
gtNextgenBranchlessLogic                         100.31%      5.69s  175.80m
gtNextgenNullSeparate                            161.98%      3.52s  283.87m
gtNextgenAllOpt                                  121.43%      4.70s  212.80m
gtCompile                                                   63.86ms    15.66
----------------------------------------------------------------------------
ltVelox_________Base                                          3.66s  273.59m
ltNextGen                                         63.21%      5.78s  172.94m
ltNextGen_______Base                                          5.80s  172.40m
ltNextgenIsBitClear                               84.04%      6.90s  144.89m
ltNextgenBranchlessSetNull                        72.87%      7.96s  125.62m
ltNextgenBranchlessLogic                          98.92%      5.86s  170.54m
ltNextgenNullSeparate                            160.44%      3.62s  276.60m
ltNextgenAllOpt                                  122.67%      4.73s  211.47m
ltCompile                                                   62.81ms    15.92
----------------------------------------------------------------------------
andVelox_________Base                                        33.54s   29.81m
andNextGen                                       372.02%      9.02s  110.91m
andNextGen_______Base                                         9.02s  110.91m
andNextgenIsBitClear                              70.74%     12.75s   78.45m
andNextgenBranchlessSetNull                       82.96%     10.87s   92.01m
andNextgenBranchlessLogic                        116.18%      7.76s  128.85m
andNextgenNullSeparate                           199.97%      4.51s  221.78m
andNextgenAllOpt                                 122.80%      7.34s  136.20m
andCompile                                                  66.94ms    14.94
----------------------------------------------------------------------------
orVelox_________Base                                         35.19s   28.42m
orNextGen                                        438.73%      8.02s  124.67m
orNextGen_______Base                                          8.07s  123.91m
orNextgenIsBitClear                               73.15%     11.03s   90.65m
orNextgenBranchlessSetNull                        78.05%     10.34s   96.71m
orNextgenBranchlessLogic                         114.90%      7.02s  142.38m
orNextgenNullSeparate                            170.23%      4.74s  210.93m
orNextgenAllOpt                                   96.96%      8.32s  120.14m
orCompile                                                   65.04ms    15.37
----------------------------------------------------------------------------
conjunctsNestedVelox_________Base                           2.74min    6.09m
conjunctsNestedNextGen                           713.07%     23.02s   43.44m
conjunctsNestedNextGen_______Base                            22.92s   43.63m
conjunctsNestedNextgenIsBitClear                  89.08%     25.73s   38.87m
conjunctsNestedNextgenBranchlessSetNull           91.30%     25.10s   39.84m
conjunctsNestedNextgenBranchlessLogic            135.07%     16.97s   58.93m
conjunctsNestedNextgenNullSeparate               373.86%      6.13s  163.13m
conjunctsNestedNextgenAllOpt                     301.28%      7.61s  131.46m
conjunctsNestedCompile                                      73.73ms    13.56
----------------------------------------------------------------------------
```

### Why are the changes needed?
try to improve performance 


### Does this PR introduce _any_ user-facing change?
No


### How was this patch tested?
benchmark tester

### Which label does this PR belong to?
INFRA