@bigPYJ1151 2023-01-16
> any thoughts on which auto-vectorization strategy to support in codegen? @winningsix @bigPYJ1151 @yma11 @jikunshang

I'm working on this. Introduce a vectorize project translator to split value and null processing without changing most expression's codgen procedure.